Update README.md
 Bone Fractures was a normal standing issue for mankind, and its classification via x-ray has always depended on human diagnostics – which may be sometimes flawed or it may have human errors and may cause the patient. To ease out this problem and as per the current level of the technology it should be simplified. The Bone Fracture is generally seen in Old Age people where the calcium level of the bone is reduced.
Computers have been shown to be valuable in every facet of human life from banking and online shopping, to Communication, Education, R&D and even in the Medical Science field. To help doctors and hospitals to examine patients well and take better care of the patients, a lot of innovative technology came into existence.
For example: X-Ray machines for identifying the bone fracture and diagnosis machine to check health issues faced by the patient. X-Rays produce a fuzzy picture of the bone and help to identify the issue, through which surgeons make an inaccurate diagnosis of the bone fractures when they use it. This undergoes various stages as Data-Preprocessing, Edge detection, Feature Extraction/Engineering and Machine Learning Algorithms that constitutes the backbone of this system, with the end goal in mind making the life of the surgeons easier.
Among the various fields that benefit from machine Learning now-a-days are sensing and medicine; this is one of the examples. As a side-note several machine learning algorithms are used for the detection of bone fracture such as Naïve Bayes, Decision Tree, Nearest Neighbors, Random Forest-Tree and SVM have been used specifically for handling the detection of the Bone Fracture detection on the dataset of 270 X-ray images. Accuracy measures for the various algorithms are in the range of around 0.62 to 0.92 which are obtained by using Naïve Bayes, Decision Tree, Nearest Neighbors, Random Forest-Tree and SVM. But Statistically accuracy of SVM has been better as per other Algorithms.